requirements.txt
requests==2.3.0
google.auth==1.5.1